+++ src/base/global.F90
@@ -73,7 +73,7 @@ module global
    !! \f$c_{\textrm{fr}} = \sqrt{v^2 + \frac{1}{2}(\max{v} - \min{v})c_{\textrm{fr}}^{\textrm{smooth}}} + \ldots\f$
    !<
    real    :: cfr_smooth
-   integer, protected :: integration_order !< Runge-Kutta time integration order (1 - 1st order, 2 - 2nd order)
+   integer(kind=4), protected :: integration_order !< Runge-Kutta time integration order (1 - 1st order, 2 - 2nd order)
    character(len=cbuff_len) :: limiter     !< type of flux limiter
    character(len=cbuff_len) :: cflcontrol  !< type of cfl control just before each sweep (possibilities: 'none', 'main', 'user')
    logical                  :: repeat_step !< repeat fluid step if cfl condition is violated (significantly increases mem usage)
