+++ src/multigrid/multigridvars.F90
@@ -115,7 +115,7 @@ contains
       implicit none
 
       class(vcycle_stats), intent(out) :: this   !< V-cycle statistics variable to be created or reset
-      integer,             intent(in)  :: size !< size of the vs structure (usually max_cycles); for nonpositive value perform reset only
+      integer(kind=4),     intent(in)  :: size !< size of the vs structure (usually max_cycles); for nonpositive value perform reset only
 
       if (size > 0) then
          if (allocated(this%factor) .or. allocated(this%time)) call die("[multigridhelpers:vcycle_stats_init] vcycle_stats already allocated.")
